// src/app/core/models/formation.ts
export interface Formation {
  idFormation: number;
  title: string;
  offreFormation: string;
  domain: string;
  theme: string;
  duree: number;
  lieu: string;
  unite: string;
}

export interface FormationFormValue {
  title: string;
  offreFormation: string;
  domain: string;
  theme: string;
  duree: number;
  lieu: string;
  unite: string;
}